refactor: ProjectDashBoardService 추가 (#316)
