+++ src/main/java/kr/mywork/interfaces/dashboard/controller/DashboardController.java
@@ -1,5 +1,14 @@
 package kr.mywork.interfaces.dashboard.controller;
 
+import java.time.LocalDate;
+import java.util.List;
+
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.Pattern;
 import kr.mywork.common.api.support.response.ApiResponse;
@@ -7,20 +16,18 @@ import kr.mywork.common.auth.components.annotation.LoginMember;
 import kr.mywork.common.auth.components.dto.LoginMemberDetail;
 import kr.mywork.domain.dashboard.service.dto.response.DashboardCountSummaryResponse;
 import kr.mywork.domain.dashboard.service.dto.response.DashboardPopularProjectsResponse;
-import kr.mywork.domain.project.service.ProjectService;
+import kr.mywork.domain.project.service.ProjectDashBoardService;
 import kr.mywork.domain.project.service.dto.request.NearDeadlineProjectRequest;
 import kr.mywork.domain.project.service.dto.response.NearDeadlineProjectResponse;
 import kr.mywork.domain.project.service.dto.response.ProjectAmountSummaryResponse;
-import kr.mywork.interfaces.dashboard.controller.dto.response.*;
+import kr.mywork.interfaces.dashboard.controller.dto.response.DashboardCountSummaryWebResponse;
+import kr.mywork.interfaces.dashboard.controller.dto.response.DashboardPopularProjectListWebResponse;
+import kr.mywork.interfaces.dashboard.controller.dto.response.DashboardPopularProjectWebResponse;
+import kr.mywork.interfaces.dashboard.controller.dto.response.NearDeadlineProjectListWebResponse;
+import kr.mywork.interfaces.dashboard.controller.dto.response.NearDeadlineProjectWebResponse;
+import kr.mywork.interfaces.dashboard.controller.dto.response.ProjectAmountChartWebResponse;
+import kr.mywork.interfaces.dashboard.controller.dto.response.ProjectAmountSummaryWebResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.time.LocalDate;
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -30,13 +37,13 @@ public class DashboardController {
 
 	private static final String AMOUNT_CHART_TYPE = "^(CHART_TYPE_WEEK|CHART_TYPE_MONTH)$";
 
-	private final ProjectService projectService;
+	private final ProjectDashBoardService projectDashboardService;
 
 	@GetMapping("/total-summary")
 	public ApiResponse<DashboardCountSummaryWebResponse> getDashboardTotalCount(
 			@LoginMember LoginMemberDetail loginMemberDetail) {
 
-		final DashboardCountSummaryResponse dashboardCountSummaryResponse = projectService.getSummaryTotalCount(loginMemberDetail);
+		final DashboardCountSummaryResponse dashboardCountSummaryResponse = projectDashboardService.getSummaryTotalCount(loginMemberDetail);
 
 		final DashboardCountSummaryWebResponse dashboardCountSummaryWebResponse = DashboardCountSummaryWebResponse.from(dashboardCountSummaryResponse);
 
@@ -48,7 +55,7 @@ public class DashboardController {
 	public    ApiResponse<DashboardPopularProjectListWebResponse> getMostPostProjectsTopFive(
 		@LoginMember final LoginMemberDetail memberDetail
 	){
-		final List<DashboardPopularProjectsResponse> popularProjects = projectService.getMostPostProjectsTopFive(memberDetail);
+		final List<DashboardPopularProjectsResponse> popularProjects = projectDashboardService.getMostPostProjectsTopFive(memberDetail);
 
 		final List<DashboardPopularProjectWebResponse> webResponse = popularProjects.stream()
 			.map(DashboardPopularProjectWebResponse::from)
@@ -75,10 +82,10 @@ public class DashboardController {
 		);
 
 		final List<NearDeadlineProjectResponse> NearDeadlineProjectResponse =
-			projectService.findNearDeadlineProjectsByLoginMember(page, nearDeadlineProjectWebRequest, today);
+			projectDashboardService.findNearDeadlineProjectsByLoginMember(page, nearDeadlineProjectWebRequest, today);
 
 		final long totalCount =
-			projectService.countNearDeadlineProjectsByLoginMember(nearDeadlineProjectWebRequest, today);
+			projectDashboardService.countNearDeadlineProjectsByLoginMember(nearDeadlineProjectWebRequest, today);
 
 		final List<NearDeadlineProjectWebResponse> nearDeadlineProjectWebResponses = NearDeadlineProjectResponse.stream()
 			.map(NearDeadlineProjectWebResponse::fromServiceResponse)
@@ -96,7 +103,7 @@ public class DashboardController {
 			@Pattern(regexp = AMOUNT_CHART_TYPE, message = "{dashboard.invalid.amount-char-type}") String chartType){
 		LocalDate today = LocalDate.now();
 		//날짜, totalCount
-		final List<ProjectAmountSummaryResponse> projectAmountSummaryList = projectService.findProjectAmountSummary(loginMemberDetail,chartType,today);
+		final List<ProjectAmountSummaryResponse> projectAmountSummaryList = projectDashboardService.findProjectAmountSummary(loginMemberDetail,chartType,today);
 
 		final List<ProjectAmountChartWebResponse> projectAmountChartWebResponses = projectAmountSummaryList.stream()
 				.map(ProjectAmountChartWebResponse::from)
